rename fxml files, move primary Controller to new controllers package, add functionality to SelectCourses view
