+++ src/ScheduleCreator/controllers/PrimaryController.java
@@ -1,13 +1,14 @@
-package ScheduleCreator;
+package ScheduleCreator.controllers;
 
 /**
  * This class serves as the controller for the primary view
  * 
  * @author Jamison Valentine, Ilyass Sfar, Nick Econopouly, Nathan Tolodzieki
  * 
- * Last Updated: 2/17/2020
+ * Last Updated: 2/20/2020
  */
 
+import ScheduleCreator.*;
 import javafx.application.Application;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
@@ -28,7 +29,7 @@ public class PrimaryController implements Initializable {
         public void changeToSelectClasses(ActionEvent _event) throws Exception {
             
             //new FXML loader and scene for new screen
-            Parent root = FXMLLoader.load(getClass().getResource("resources/select_classes.fxml"));
+            Parent root = FXMLLoader.load(getClass().getResource("/ScheduleCreator/resources/select_courses.fxml"));
             Scene classViewScene = new Scene(root);
             
             //Get information from primary stage
@@ -37,20 +38,9 @@ public class PrimaryController implements Initializable {
             window.show();
         }
         
-        public void backToPrimary(ActionEvent _event) throws Exception {
-            
-            Parent root = FXMLLoader.load(getClass().getResource("resources/primary.fxml"));
-            Scene primaryScene = new Scene(root);
-            
-            //Get information from primary stage
-            Stage window = (Stage)((Node)_event.getSource()).getScene().getWindow();
-            window.setScene(primaryScene);
-            window.show();
-        }
-        
         public void changeToRegistrationScreen(ActionEvent _event) throws Exception {
             
-            Parent root = FXMLLoader.load(getClass().getResource("resources/registration_screen.fxml"));
+            Parent root = FXMLLoader.load(getClass().getResource("/ScheduleCreator/resources/registration_screen.fxml"));
             Scene scene = new Scene(root);
             
             Stage stage = (Stage)((Node)_event.getSource()).getScene().getWindow();
