--- src/ScheduleCreator/resources/select_classes.fxml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ChoiceBox?>
-<?import javafx.scene.control.ListView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.StackPane?>
-
-<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="600.0" stylesheets="@style.css" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ScheduleCreator.PrimaryController">
-   <children>
-       <Button onAction="#backToPrimary" text="Back" />
-      <StackPane layoutX="200.0" layoutY="131.0" prefHeight="150.0" prefWidth="200.0">
-         <children>
-            <ChoiceBox prefWidth="150.0" />
-            <ListView /> 
-
-         </children>
-      </StackPane>
-   </children>
-</AnchorPane>
