+++ src/ScheduleCreator/resources/select_courses.fxml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.shape.*?>
+<?import java.lang.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ChoiceBox?>
+<?import javafx.scene.control.ListView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.StackPane?>
+
+<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="600.0" stylesheets="@style.css" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ScheduleCreator.controllers.CoursesController">
+   <children>
+       <Button layoutX="14.0" layoutY="22.0" onAction="#backToPrimary" text="Back" />
+       <Button layoutX="71.0" layoutY="22.0" onAction="#addCourse" text="Add Course" />
+      <StackPane layoutX="200.0" layoutY="131.0" prefHeight="150.0" prefWidth="200.0" />
+      <AnchorPane layoutX="386.0" layoutY="16.0" prefHeight="368.0" prefWidth="200.0">
+         <children>
+            <Label layoutX="52.0" layoutY="19.0" text="Find Courses: " />
+            <TextField layoutX="26.0" layoutY="45.0" />
+         </children>
+      </AnchorPane>
+      <ScrollPane layoutX="34.0" layoutY="124.0" maxHeight="1000.0" prefViewportHeight="252.0" prefViewportWidth="315.0">
+         <content>
+            <GridPane fx:id="courses" hgap="5.0" layoutX="34.0" layoutY="124.0" prefHeight="237.0" prefWidth="300.0" vgap="5.0">
+               <columnConstraints>
+                  <ColumnConstraints />
+               </columnConstraints>
+               <rowConstraints>
+                  <RowConstraints />
+                  <RowConstraints />
+               </rowConstraints>
+               <children>
+
+               </children></GridPane>
+         </content>
+      </ScrollPane>
+   </children>
+</AnchorPane>
